Rename 'tk_maya_plugin_basic' to 'tk_maya_basic'.
