+++ plugins/basic/python/tk_maya_basic/__init__.py
