+++ plugins/basic/python/tk_maya_basic/plugin_logging.py
